refactor(wiki): extract the fs/YAML server read-engine into @mulmoclaude/core/wiki/server
Phase 1.5 of bringing Wiki to MulmoTerminal. Phase 1 shared the pure
engine but left the filesystem/YAML read layer host-side, which would
have forced MT to reimplement slug resolution, the read loops behind
graph/lint, and frontmatter parsing — diverging readers over the shared
data/wiki/ files. This promotes that engine into core (mirroring
collection/server) so both hosts read identically.

@mulmoclaude/core/wiki/server now also exports the workspace-injected
read-engine:
- getPageIndex, resolvePagePath (+ pickFuzzyMatch)
- readWikiIndex / readWikiPage / readWikiLog
- loadWikiGraph, collectLintIssues
- parseFrontmatter / parseFrontmatterTags (read-side js-yaml, FAILSAFE_SCHEMA)
- wikiDirs (canonical data/wiki/{index.md,pages,log.md} layout)

Path layout + the fs/YAML deps live in core; the host route
(server/api/routes/wiki.ts) is now a thin Express shell that shapes the
core engine's output into the canvas response envelope. The write side
(writeWikiPage / snapshots) stays host-side until a host needs it.

Also adds a lightweight ./wiki/paths subpath (wikiSlugFromAbsPath /
wikiDirs, no js-yaml) so the committed hooks sidecar bundle imports the
path helper without inlining the whole YAML parser — wikiSnapshot.ts and
wiki-pages/io.ts use it.

core 0.4.0 -> 0.5.0 (additive exports + js-yaml dep); consumers
(packages/mulmoclaude, collection-plugin) bumped to ^0.5.0. The pure
resolver/pageIndex unit tests move into packages/core/test/wiki.
