+++ packages/core/src/wiki/server/index.ts
@@ -1,6 +1,9 @@
-// Node-only wiki helpers (need `node:path`) — kept off the
-// browser-safe `@mulmoclaude/core/wiki` surface. Today: the
-// abs-path → slug resolver used by the host's wiki-write
-// chokepoint and the PostToolUse snapshot hook.
+// Node-only wiki helpers — kept off the browser-safe
+// `@mulmoclaude/core/wiki` surface. The workspace-injected read engine
+// (resolve / index / page / log / graph / lint) plus the abs-path→slug
+// resolver used by the host's write chokepoint and snapshot hook.
 
-export { wikiSlugFromAbsPath } from "./paths.js";
+export { wikiSlugFromAbsPath, wikiDirs } from "./paths.js";
+export { getPageIndex, __resetPageIndexCache, type PageIndex } from "./pageIndex.js";
+export { parseFrontmatter, parseFrontmatterTags } from "./frontmatter.js";
+export { pickFuzzyMatch, resolvePagePath, readWikiIndex, readWikiLog, readWikiPage, loadWikiGraph, collectLintIssues, type WikiPageRead } from "./engine.js";
