+++ packages/core/vite.config.ts
@@ -25,6 +25,10 @@ export default defineConfig({
         "collection/registry/server/index": "src/collection/registry/server/index.ts",
         "wiki/index": "src/wiki/index.ts",
         "wiki/server/index": "src/wiki/server/index.ts",
+        // Lightweight path helpers as their own entry (no js-yaml / engine)
+        // so the hook sidecar can import `wikiSlugFromAbsPath` without
+        // bundling the whole read-engine. Mirrors `collection/paths`.
+        "wiki/paths": "src/wiki/server/paths.ts",
         "feeds/index": "src/feeds/index.ts",
         "feeds/server/index": "src/feeds/server/index.ts",
         "feeds/paths": "src/feeds/paths.ts",
@@ -41,7 +45,7 @@ export default defineConfig({
       fileName: (format, entryName) => `${entryName}.${format === "es" ? "js" : "cjs"}`,
     },
     rollupOptions: {
-      external: [/^node:/, /^@receptron\//, "zod", "gui-chat-protocol", "fast-xml-parser"],
+      external: [/^node:/, /^@receptron\//, "zod", "gui-chat-protocol", "fast-xml-parser", "js-yaml"],
       output: { exports: "named" },
     },
     minify: false,
