+++ packages/mulmoclaude/package.json
@@ -36,7 +36,7 @@
     "@mulmoclaude/accounting-plugin": "^0.2.0",
     "@mulmoclaude/chart-plugin": "^0.1.1",
     "@mulmoclaude/collection-plugin": "^0.5.18",
-    "@mulmoclaude/core": "^0.4.0",
+    "@mulmoclaude/core": "^0.5.0",
     "@mulmoclaude/form-plugin": "^0.1.0",
     "@mulmoclaude/html-plugin": "^0.2.0",
     "@mulmoclaude/markdown-plugin": "^0.1.2",
