+++ server/workspace/hooks/handlers/wikiSnapshot.ts
@@ -6,7 +6,7 @@
 // carries a single PostToolUse entry instead of one per hook source.
 
 import path from "node:path";
-import { wikiSlugFromAbsPath } from "@mulmoclaude/core/wiki/server";
+import { wikiSlugFromAbsPath } from "@mulmoclaude/core/wiki/paths";
 import { buildAuthPost, safePost } from "../shared/sidecar.js";
 import type { HookPayload } from "../shared/stdin.js";
 import { extractFilePath, extractSessionId, extractToolName } from "../shared/stdin.js";
