+++ server/workspace/wiki-pages/io.ts
@@ -24,7 +24,7 @@ import { readTextSafe } from "../../utils/files/safe.js";
 import { writeFileAtomic } from "../../utils/files/atomic.js";
 import { mergeFrontmatter, parseFrontmatter, serializeWithFrontmatter } from "../../utils/markdown/frontmatter.js";
 import { isSafeSlug } from "@mulmoclaude/core/wiki";
-import { wikiSlugFromAbsPath } from "@mulmoclaude/core/wiki/server";
+import { wikiSlugFromAbsPath } from "@mulmoclaude/core/wiki/paths";
 import { workspacePath as defaultWorkspacePath } from "../workspace.js";
 import { WORKSPACE_DIRS } from "../paths.js";
 import { appendSnapshot } from "./snapshot.js";
